fix compile error

+++ src/zip.rs
@@ -105,7 +105,7 @@ pub fn write_yarn_zip(
 
             match header.entry_type() {
                 tar::EntryType::Regular => {
-                    let (src, must_flush) = if entry.size() >= BIG_FILE_THRESHOLD {
+                    let (src, must_flush) = if entry.size() >= (BIG_FILE_THRESHOLD as u64) {
                         // The file is >= BIG_FILE_THRESHOLD, pass it as a streaming reader to avoid buffering it,
                         // and immediately close the archive to force reading while the input tar
                         // stream is still at the position of that Entry.
